tests/wpt/web-platform-tests/pointerevents/parsing/touch-action-valid.html

<!DOCTYPE html>
<html>
<head>
<title>Pointer Events: parsing touch-action with valid values</title>
<link rel="help" href="https://w3c.github.io/pointerevents/#the-touch-action-css-property">
<meta name="assert" content="touch-action supports the full grammar 'auto | none | [ pan-x || pan-y ] | manipulation'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("touch-action", "auto");
test_valid_value("touch-action", "none");
test_valid_value("touch-action", "manipulation");

// [ pan-x || pan-y ]
test_valid_value("touch-action", "pan-x");
test_valid_value("touch-action", "pan-y");
test_valid_value("touch-action", "pan-y pan-x", "pan-x pan-y");
</script>
</body>
</html>
